+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 Version numbers correspond to `package.json` version
 
+# 2.6.7-2.6.9 (12/16/2020)
+- Solr - Disable highlight unless enabled in the fieldSpec
+
+# 2.6.6 (12/09/2020)
+- Allow full query override for ES explain other
+
+# 2.6.4 (11/30/2020)
+- Allow override of defType for explainOther in solr searcher
+
+# 2.6.3 (10/21/2020)
+- Fixed issue with documents having no title causing splainer to blow up.
+
+# 2.6.2 (10/05/2020)
+- Lots of work on breaking out highlighting into a better structure, after fighting with little fixes for a while.
+
 # 2.5.9 (04/27/2020)
 - Highlighting on dates and integers in Solr causes an error.  Work around is to append `hl.method=unified` to calls to Solr.  https://github.com/o19s/splainer-search/pull/84
 - a common pattern in Solr schemas is to normalize fields with dots: `foo.bar`, however if you have a array or dictionary of JSON, we want to navigate that.  Now we check if the key exists with a dot in it, if not, we use that as a selector to pluck out the data in the nested JSON that we need.  https://github.com/o19s/splainer-search/pull/83